refactored codec tests
