+++ src/test/java/com/amazon/opendistroforelasticsearch/knn/index/codec/KNN80Codec/KNN80CodecTest.java
@@ -1,0 +1,29 @@
+/*
+ *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License").
+ *   You may not use this file except in compliance with the License.
+ *   A copy of the License is located at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file. This file is distributed
+ *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *   express or implied. See the License for the specific language governing
+ *   permissions and limitations under the License.
+ */
+
+package com.amazon.opendistroforelasticsearch.knn.index.codec.KNN80Codec;
+
+import com.amazon.opendistroforelasticsearch.knn.index.codec.KNNCodecTestCase;
+
+public class  KNN80CodecTest extends KNNCodecTestCase {
+
+    public void testFooter() throws Exception {
+        testFooter(new KNN80Codec());
+    }
+
+    public void testMultiFieldsKnnIndex() throws Exception {
+        testMultiFieldsKnnIndex(new KNN80Codec());
+    }
+}
